requirements.txt
numpy>=1.26
pandas>=2.2
scikit-learn>=1.3
matplotlib>=3.8
seaborn>=0.13
jupyterlab>=4.0  # optional - useful if using notebooks
python-docx
reportlab
